fixLiveDataVariable

+++ src/fms_main/src/_fms_lanfeng.h
@@ -109,13 +109,13 @@
      }
      
 
-     uint8_t readPumpState(uint16_t registerAddress){
+     uint32_t readPumpState(uint16_t registerAddress){
         uint8_t result = _node.readHoldingRegisters(registerAddress,1);
         if(result == _node.ku8MBSuccess) return _node.getResponseBuffer(0);
         else result;
      }
 
-     uint8_t readLiveData(uint16_t registerAddress){
+     uint32_t readLiveData(uint16_t registerAddress){
       uint8_t result = _node.readHoldingRegisters(registerAddress,1);
       if(result == _node.ku8MBSuccess) return _node.getResponseBuffer(0);
       else return result;
